[BROWSEUI] [OLEAUT32] [DDK] [PSDK] [STLPORT] [CPPRT] Port the vs2015 fixes to also work on x64
[HHCOMP]
Fix compilation with VS in x64 mode

svn path=/trunk/; revision=68791

+++ reactos/lib/3rdparty/stlport/test/eh/nc_alloc.cpp
@@ -265,9 +265,9 @@ void _STLP_CALL operator delete(void* s)
 }
 
 #if defined (EH_DELETE_HAS_THROW_SPEC)
-void _STLP_CALL operator delete(void* s, unsigned int) throw()
+void _STLP_CALL operator delete(void* s, uintptr_t) throw()
 #else
-void _STLP_CALL operator delete(void* s, unsigned int)
+void _STLP_CALL operator delete(void* s, uintptr_t)
 #endif
 {
 	::operator delete(s);
